split files to match data
